refactor: camel case apps
