+++ now/dataclasses.py
@@ -1,0 +1,27 @@
+from dataclasses import dataclass
+from typing import Optional
+
+from now.constants import Apps, DatasetTypes, Modalities, Qualities
+
+
+@dataclass
+class UserInput:
+    output_modality: Optional[Modalities] = None
+    app: Optional[Apps] = None
+
+    # data related
+    data: Optional[str] = None
+    is_custom_dataset: Optional[bool] = None
+
+    custom_dataset_type: Optional[DatasetTypes] = None
+    dataset_secret: Optional[str] = None
+    dataset_url: Optional[str] = None
+    dataset_path: Optional[str] = None
+
+    # model related
+    quality: Optional[Qualities] = Qualities.MEDIUM
+
+    # cluster related
+    cluster: Optional[str] = None
+    create_new_cluster: Optional[bool] = False
+    deployment_type: Optional[str] = None
